mcad/requirements.txt
solidpython==1.1.1
black>=21.11b1
regex>=2021.4.4
